requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.354
openai==1.3.5
pydantic==2.5.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.1.0
python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
